Changed waveform id from uint8_t to uint32_t.

+++ helper/satellite-frame-conf.cc
@@ -61,7 +61,7 @@ SatTimeSlotConf::SatTimeSlotConf ()
   NS_ASSERT (false);
 }
 
-SatTimeSlotConf::SatTimeSlotConf (double startTime_s, double duration_s, uint8_t waveFormId, uint32_t frameCarrierId)
+SatTimeSlotConf::SatTimeSlotConf (double startTime_s, double duration_s, uint32_t waveFormId, uint32_t frameCarrierId)
   : m_startTime_s (startTime_s),
     m_duration_s (duration_s),
     m_waveFormId (waveFormId),
